requirements.txt
# Core Google ADK and AI Platform dependencies
google-adk>=1.0.0
google-cloud-aiplatform>=1.42.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0

# HTTP requests and web scraping alternatives
requests>=2.31.0
beautifulsoup4>=4.12.2

# Email functionality
sendgrid>=6.10.0

# Environment management
python-dotenv>=1.0.0

# Note: Virtual environment and development tools not included in production requirements